requirements-gpu.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
coveralls==1.11.1
dill==0.3.1.1
docopt==0.6.2
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
more-itertools==8.2.0
multiprocess==0.70.9
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
p-tqdm==1.3
packaging==20.1
pandas==1.0.1
pathos==0.2.5
pluggy==0.13.1
pox==0.2.7
ppft==1.6.6.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
scikit-learn==0.22.2
six==1.14.0
tensorboard==2.1.0
tensorflow-gpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.43.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.0
zipp==3.0.0
